refactor: extract management loading from AgentWorkspace
